third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/scroll-state/at-container-scroll-direction-parsing.html

<!DOCTYPE html>
<title>CSS Conditional Test: @container scroll-direction query parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/6400">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<div style="container-name:name;container-type:scroll-state">
  <main id="cq-main"></main>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  test_cq_condition_known('scroll-state(scroll-direction)');
  test_cq_condition_known('scroll-state(scroll-direction: none)');
  test_cq_condition_known('scroll-state(scroll-direction: any)');
  test_cq_condition_known('scroll-state(scroll-direction: top)');
  test_cq_condition_known('scroll-state(scroll-direction: left)');
  test_cq_condition_known('scroll-state(scroll-direction: bottom)');
  test_cq_condition_known('scroll-state(scroll-direction: right)');
  test_cq_condition_known('scroll-state(scroll-direction: inline-start)');
  test_cq_condition_known('scroll-state(scroll-direction: inline-end)');
  test_cq_condition_known('scroll-state(scroll-direction: block-start)');
  test_cq_condition_known('scroll-state(scroll-direction: block-end)');
  test_cq_condition_known('scroll-state(scroll-direction: block)');
  test_cq_condition_known('scroll-state(scroll-direction: x)');
  test_cq_condition_known('scroll-state(scroll-direction: y)');
  test_cq_condition_known('scroll-state(scroll-direction: inline)');
  test_cq_condition_known('(scroll-state(scroll-direction: inline-end))');
  test_cq_condition_known('scroll-state((scroll-direction: left))');
  test_cq_condition_known('scroll-state(not ((scroll-direction: bottom) and (scroll-direction: right)))');
  test_cq_condition_known('scroll-state((scroll-direction: left) or (scroll-direction: top))');

  test_cq_condition_unknown('scroll-state(scroll-direction: auto)');
  test_cq_condition_unknown('scroll-state(scroll-direction: true)');
  test_cq_condition_unknown('scroll-state(style(scroll-direction: left))');
  test_cq_condition_unknown('style(scroll-state(scroll-direction: left))');
  test_cq_condition_unknown('scroll-state(scroll-direction:)');
</script>
